dernières retouches

+++ app/controller/comments.controller.js
@@ -10,7 +10,6 @@ exports.getAll = async (req, res) => {
    await jwt.verifyToken(req, res);
    try {
       const publi = await Publication.findByPk(req.params.id);
-      console.log(publi);
       if (publi === null) {
          res.status(500).send({message: `Publication n°${req.params.id} inexistante`}); 
       } else {
@@ -28,14 +27,18 @@ exports.getById = async (req, res) => {
    await jwt.verifyToken(req, res);
    try {
       const publi = await Publication.findByPk(req.params.id);
-      const idCom = req.params.idComment;
-      const comment = await Comment.findByPk(idCom);
-      const ship = await publi.hasComment(comment);
-      if (ship) {
-         let result = await Comment.findOne({where: {id: idCom}}); // = .findByPk(id);
-         res.status(200).json(result);
+      if (publi === null) {
+         res.status(500).send({message: `Publication n°${req.params.id} inexistante`}); 
       } else {
-         res.status(400).json({message: "Aucun Commentaire id = " + id});
+         const idCom = req.params.idComment;
+         const comment = await Comment.findByPk(idCom);
+         const ship = await publi.hasComment(comment);
+         if (ship) {
+            let result = await Comment.findOne({where: {id: idCom}}); // = .findByPk(id);
+            res.status(200).json(result);
+         } else {
+            res.status(400).json({message: "Aucun commentaire trouvé"});
+         }
       }
    } catch (err) {
       res.status(500).send({message: err});
@@ -45,27 +48,29 @@ exports.getById = async (req, res) => {
  // POST /publications/:id/comments
  exports.create = async (req, res) => {
    let userId = await jwt.verifyToken(req, res);
-   const user = await User.findOne({where: {id: userId}});
    if (!req.body) {
       res.status(400).send({message: "Tous les champs sont obligatoires!"});
       return;
    }
    try {
-      const idPubli = req.params.id;
-      let publi = await Publication.findByPk(idPubli);
-      if (user.dataValues.type == 1) { // Type 1 pour Professeur
-         let comment = await publi.createComment(req.body);
-         let teacher = await user.getTeacher();
-         await teacher.addComment(comment);
-         res.status(200).json(comment);   
-      } else if (user.dataValues.type == 2) { // Type 2 pour Etudiant
-         let comment = await publi.createComment(req.body);
-         let student = await user.getStudent();
-         await student.addComment(comment);
-         res.status(200).json(comment);  
+      let publi = await Publication.findByPk(req.params.id);
+      if (publi === null) {
+         res.status(500).send({message: `Publication n°${req.params.id} inexistante`}); 
       } else {
-         res.status(401).json({message: "Vous ne disposez pas des droits."});
-      }      
+         const user = await User.findOne({where: {id: userId}});
+         let comment = await publi.createComment(req.body);
+         if (user.dataValues.type == 1) { // Type 1 pour Professeur
+            let teacher = await user.getTeacher();
+            await teacher.addComment(comment);
+            res.status(200).json(comment);   
+         } else if (user.dataValues.type == 2) { // Type 2 pour Etudiant
+            let student = await user.getStudent();
+            await student.addComment(comment);
+            res.status(200).json(comment);  
+         } else {
+            res.status(401).json({message: "Vous ne disposez pas des droits."});
+         }  
+      }    
    } catch (err) {
       res.status(500).send({ message: err.message || "Erreur"});
    }
@@ -79,30 +84,44 @@ exports.getById = async (req, res) => {
       return;
    }
    try {
-      let comment = await Comment.findByPk(req.params.idComment);
-      let userToken = await User.findOne({where: {id: userId}});
-      if (userToken.dataValues.type == 1) { // Type 1 pour Professeur
-         let teacher = await userToken.getTeacher();
-         let ship = await teacher.hasComment(comment);
-         if (ship) {
-            await comment.update(req.body, {where: {id: req.params.idComment}});
-            let newCom = new CommentClass(req.body);
-            newCom.id = parseInt(req.params.idComment);
-            res.status(200).json({ comment : newCom }); 
-         } 
-      } else if (userToken.dataValues.type == 2) { // Type 2 pour Etudiant
-         let student = await userToken.getStudent();
-         let ship = await student.hasComment(comment);
-         if (ship) {
-            await comment.update(req.body, {where: {id: req.params.idComment}});
-            let newCom = new CommentClass(req.body);
-            newCom.id = parseInt(req.params.idComment);
-            res.status(200).json({ comment : newCom }); 
-         } 
+      let publi = await Publication.findByPk(req.params.id);
+      if (publi === null) {
+         res.status(500).send({message: `Publication n°${req.params.id} inexistante`}); 
       } else {
-         res.status(401).json({message: "Vous ne disposez pas des droits."});
-      }
+         let comment = await Comment.findByPk(req.params.idComment);
+         if (comment === null) {
+            res.status(500).json({message: `Commentaire inexistant`}); 
+         } else {
 
+            let userToken = await User.findOne({where: {id: userId}});
+            if (userToken.dataValues.type == 1) { // Type 1 pour Professeur
+               let teacher = await userToken.getTeacher();
+               let ship = await teacher.hasComment(comment);
+               if (ship) {
+                  await comment.update(req.body, {where: {id: req.params.idComment}});
+                  let newCom = new CommentClass(req.body);
+                  newCom.id = parseInt(req.params.idComment);
+                  res.status(200).json({ comment : newCom }); 
+               } else {
+                  res.status(401).json({message: "Vous ne disposez pas des droits."});
+               }
+            } else if (userToken.dataValues.type == 2) { // Type 2 pour Etudiant
+               let student = await userToken.getStudent();
+               let ship = await student.hasComment(comment);
+               if (ship) {
+                  await comment.update(req.body, {where: {id: req.params.idComment}});
+                  let newCom = new CommentClass(req.body);
+                  newCom.id = parseInt(req.params.idComment);
+                  res.status(200).json({ comment : newCom }); 
+               } else {
+                  res.status(401).json({message: "Vous ne disposez pas des droits."});
+               }
+            } else if (userToken.dataValues.type == 3) {
+               res.status(401).json({message: "Vous ne disposez pas des droits."});
+            }
+
+         }
+      }
    } catch (err) {
       res.status(500).send({ message: err.message || "Erreur"});
    }
@@ -111,35 +130,47 @@ exports.getById = async (req, res) => {
 // DELETE /publications/:id/comments/:idComment
 exports.remove = async (req, res) => {
    let userId = await jwt.verifyToken(req, res);
-
    try {
-      let comment = await Comment.findByPk(req.params.idComment);
-      let userToken = await User.findOne({where: {id: userId}});
-      if (userToken.dataValues.type == 1) { // Type 1 pour Professeur
-         let teacher = await userToken.getTeacher();
-         let ship = await teacher.hasComment(comment);
-         if (ship) {
-            await Comment.destroy({where: { id: req.params.idComment }});
-            let result = await Comment.findByPk(req.params.idComment);
-            if (result === null) {
-               res.status(200).send({ message: "Le commentaire a été supprimé!"});
-            }
-         }
-      } else if (userToken.dataValues.type == 2) { // Type 2 pour Etudiant
-         let student = await userToken.getStudent();
-         let ship = await student.hasComment(comment);
-         if (ship) {
-            await Comment.destroy({where: { id: req.params.idComment }});
-            let result = await Comment.findByPk(req.params.idComment);
-            if (result === null) {
-               res.status(200).send({ message: "Le commentaire a été supprimé!"});
-            }
-         }
+      let publi = await Publication.findByPk(req.params.id);
+      if (publi === null) {
+         res.status(500).json({message: `Publication n°${req.params.id} inexistante`}); 
       } else {
-         res.status(401).json({message: "Vous ne disposez pas des droits."});
-      }
+         let comment = await Comment.findByPk(req.params.idComment);
+         if (comment === null) {
+            res.status(500).json({message: `Commentaire inexistant`}); 
+         } else {
 
+            let userToken = await User.findOne({where: {id: userId}});
+            if (userToken.dataValues.type == 1) { // Type 1 pour Professeur
+               let teacher = await userToken.getTeacher();
+               let ship = await teacher.hasComment(comment);
+               if (ship) {
+                  await Comment.destroy({where: { id: req.params.idComment }});
+                  let result = await Comment.findByPk(req.params.idComment);
+                  if (result === null) {
+                     res.status(200).json({ message: "Le commentaire a été supprimé!"});
+                  }
+               } else {
+                  res.status(401).json({message: "Vous ne disposez pas des droits."});
+               }
+            } else if (userToken.dataValues.type == 2) { // Type 2 pour Etudiant
+               let student = await userToken.getStudent();
+               let ship = await student.hasComment(comment);
+               if (ship) {
+                  await Comment.destroy({where: { id: req.params.idComment }});
+                  let result = await Comment.findByPk(req.params.idComment);
+                  if (result === null) {
+                     res.status(200).json({ message: "Le commentaire a été supprimé!"});
+                  }
+               } else {
+                  res.status(401).json({message: "Vous ne disposez pas des droits."});
+               }
+            } else if (userToken.dataValues.type == 3) {
+               res.status(401).json({message: "Vous ne disposez pas des droits."});
+            }
+         }
+      }
    } catch (err) {
-      res.status(500).send({ message: err.message || "Erreur"});
+      res.status(500).json({ message: err.message || "Erreur"});
    }
 }
